requirements.txt
langchain<0.2
langchainhub
langchain-community
langchain_core
langchain_openai
langchain_text_splitters
openai
streamlit
pypdf
fastembed
docarray
validators
python-dotenv
google-api-python-client
pydantic_settings